fix(viewer): split dashboard into multiple components and move second to topic tree
