+++ src/assets/scss/mixins.scss
@@ -191,3 +191,21 @@
     font-weight: bold;
   }
 }
+
+@mixin color-picker-item {
+  .color-picker-item {
+    .el-form-item__content {
+      display: flex !important;
+      align-items: center !important;
+      line-height: normal !important;
+      height: 43px !important;
+      justify-content: flex-start !important;
+    }
+
+    .el-color-picker {
+      vertical-align: middle !important;
+      margin: 0 !important;
+      align-self: center !important;
+    }
+  }
+}
